+++ lib/asimov-snapshot/src/lib.rs
@@ -12,13 +12,3 @@ pub use snapshot::*;
 
 #[cfg(feature = "std")]
 pub mod storage;
-
-#[derive(Clone, Debug, PartialEq, Eq, bon::Builder)]
-pub struct Snapshot {
-    #[builder(into)]
-    pub url: std::string::String,
-    #[builder(into)]
-    pub data: std::vec::Vec<u8>,
-    pub start_timestamp: jiff::Timestamp,
-    pub end_timestamp: Option<jiff::Timestamp>,
-}
